pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "bunyang-longtail"
version = "0.1.0"
description = "청약 롱테일 주제/SEO 프롬프트 관리 도구"
requires-python = ">=3.11"
dependencies = [
    "Pillow>=10",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
